# Conflicts:
#	src/main/java/org/example/gameconnectbackend/repositories/UserRepository.java

+++ src/main/java/org/example/gameconnectbackend/repositories/UserRepository.java
@@ -3,6 +3,5 @@ package org.example.gameconnectbackend.repositories;
 import org.example.gameconnectbackend.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
